+++ src/Symfony/Bundle/WebServerBundle/Command/ServerCommand.php
@@ -1,0 +1,30 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Bundle\WebServerBundle\Command;
+
+use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
+
+/**
+ * Base methods for commands related to a local web server.
+ *
+ * @author Christian Flothmann <christian.flothmann@xabbuh.de>
+ */
+abstract class ServerCommand extends ContainerAwareCommand
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function isEnabled()
+    {
+        return !defined('HHVM_VERSION') && parent::isEnabled();
+    }
+}