+++ src/Symfony/Bundle/WebServerBundle/Command/ServerRunCommand.php
@@ -1,0 +1,117 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Bundle\WebServerBundle\Command;
+
+use Symfony\Bundle\WebServerBundle\WebServer;
+use Symfony\Bundle\WebServerBundle\WebServerConfig;
+use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Output\ConsoleOutputInterface;
+use Symfony\Component\Console\Style\SymfonyStyle;
+
+/**
+ * Runs Symfony application using a local web server.
+ *
+ * @author Michał Pipa <michal.pipa.xsolve@gmail.com>
+ */
+class ServerRunCommand extends ServerCommand
+{
+    /**
+     * {@inheritdoc}
+     */
+    protected function configure()
+    {
+        $this
+            ->setDefinition(array(
+                new InputArgument('addressport', InputArgument::OPTIONAL, 'The address to listen to (can be address:port, address, or port)'),
+                new InputOption('docroot', 'd', InputOption::VALUE_REQUIRED, 'Document root'),
+                new InputOption('router', 'r', InputOption::VALUE_REQUIRED, 'Path to custom router script'),
+            ))
+            ->setName('server:run')
+            ->setDescription('Runs a local web server')
+            ->setHelp(<<<'EOF'
+The <info>%command.name%</info> runs a local web server:
+
+  <info>%command.full_name%</info>
+
+Change the default address and port by passing them as an argument:
+
+  <info>%command.full_name% 127.0.0.1:8080</info>
+
+Use the <info>--docroot</info> option to change the default docroot directory:
+
+  <info>%command.full_name% --docroot=htdocs/</info>
+
+Specify your own router script via the <info>--router</info> option:
+
+  <info>%command.full_name% --router=app/config/router.php</info>
+
+See also: http://www.php.net/manual/en/features.commandline.webserver.php
+EOF
+            )
+        ;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function execute(InputInterface $input, OutputInterface $output)
+    {
+        $io = new SymfonyStyle($input, $output);
+
+        if (null === $documentRoot = $input->getOption('docroot')) {
+            $documentRoot = $this->getContainer()->getParameter('kernel.root_dir').'/../web';
+        }
+
+        if (!is_dir($documentRoot)) {
+            $io->error(sprintf('The document root directory "%s" does not exist.', $documentRoot));
+
+            return 1;
+        }
+
+        $env = $this->getContainer()->getParameter('kernel.environment');
+        if ('prod' === $env) {
+            $io->error('Running this server in production environment is NOT recommended!');
+        }
+
+        $callback = null;
+        $disableOutput = false;
+        if ($output->isQuiet()) {
+            $disableOutput = true;
+        } else {
+            $callback = function ($type, $buffer) use ($output) {
+                if (Process::ERR === $type && $output instanceof ConsoleOutputInterface) {
+                    $output = $output->getErrorOutput();
+                }
+                $output->write($buffer, false, OutputInterface::OUTPUT_RAW);
+            };
+        }
+
+        try {
+            $server = new WebServer();
+            $config = new WebServerConfig($documentRoot, $env, $input->getArgument('addressport'), $input->getOption('router'));
+
+            $io->success(sprintf('Server listening on http://%s', $config->getAddress()));
+            $io->comment('Quit the server with CONTROL-C.');
+
+            $exitCode = $server->run($config, $disableOutput, $callback);
+        } catch (\Exception $e) {
+            $io->error($e->getMessage());
+
+            return 1;
+        }
+
+        return $exitCode;
+    }
+}