+++ src/Symfony/Bundle/WebServerBundle/Resources/router.php
@@ -30,13 +30,15 @@ if (is_file($_SERVER['DOCUMENT_ROOT'].DIRECTORY_SEPARATOR.$_SERVER['SCRIPT_NAME'
     return false;
 }
 
+$script = getenv('APP_FRONT_CONTROLLER') ?: 'index.php';
+
 $_SERVER = array_merge($_SERVER, $_ENV);
-$_SERVER['SCRIPT_FILENAME'] = $_SERVER['DOCUMENT_ROOT'].DIRECTORY_SEPARATOR.'app_dev.php';
+$_SERVER['SCRIPT_FILENAME'] = $_SERVER['DOCUMENT_ROOT'].DIRECTORY_SEPARATOR.$script;
 
 // Since we are rewriting to app_dev.php, adjust SCRIPT_NAME and PHP_SELF accordingly
-$_SERVER['SCRIPT_NAME'] = DIRECTORY_SEPARATOR.'app_dev.php';
-$_SERVER['PHP_SELF'] = DIRECTORY_SEPARATOR.'app_dev.php';
+$_SERVER['SCRIPT_NAME'] = DIRECTORY_SEPARATOR.$script;
+$_SERVER['PHP_SELF'] = DIRECTORY_SEPARATOR.$script;
 
-require 'app_dev.php';
+require $script;
 
 error_log(sprintf('%s:%d [%d]: %s', $_SERVER['REMOTE_ADDR'], $_SERVER['REMOTE_PORT'], http_response_code(), $_SERVER['REQUEST_URI']), 4);